+++ shared/src/index.ts
@@ -10,5 +10,8 @@ export * from "./plugins/pssh-extractor/pssh-extractor.js";
 export * from "./plugins/webvtt-parser/webvtt-parser.js";
 export * from "./report.js";
 export * from "./utils/array-utils.js";
+export * from "./utils/float-utils.js";
+export * from "./utils/logger.js";
 export type { Emsg } from "./utils/mp4/types.js";
+export * from "./utils/time-range.js";
 export * from "./utils/url.js";